requirements.txt
boto3==1.42.40
Flask==3.1.2
pandas==3.0.0
psycopg2==2.9.9
python-dotenv==1.2.1
Requests==2.32.5
twilio==9.10.0
